+++ wthc.rb
@@ -1,17 +1,12 @@
 # Author: BeRogue01
-# License: Free yo, like air and water ... so far ...
 # Date: 10/12/2021
 #
 # Love everybody,but never sell your sword.  ~ Paulo Coelho
 #
 require 'tty-screen'
-load './core/common.rb'
+load './lib/common.rb'
 
-threaded = true
-numcols = TTY::Screen.cols
-$page = last_page = 1
-
-$app = app = Core.new(
+app = Core.new(
   :config_file => $options.config_file
 )
 app.clear
@@ -23,23 +18,26 @@ if app.newer_config_version?
   sleep(5)
 end
 
-#clear
-#console_header().each {|o| clear_line; puts o; }
+app.load_templates
+app.start
 
-run_mods = app.init_wth_modules(app.config['modules'])
-app.webserver_start
-app.init_key_reader
+puts "Loading data into modules..."
 
-$CFG = app.load_config
+numcols = TTY::Screen.cols
+# TODO : Fix $page global, move page state to app
+$page = last_page = 1
+last_run = Time.now - 10
+threaded = true # Turn into config option
 
-# Main Loop
+Process.daemon(true, true) if !app.console_out && !app.os.windows?
+
+# Endless good times.
 # - add web logs to web channel
 # - setup screen stuff
 # - iterate thru the modules
-puts "Loading data into modules..."
-app.load_templates
-
-last_run = Time.now - 10
+# TODO: Integrate loop into app lib and add into start
+# TODO: Decouple loop timer with running modules so we can
+#       print old screen on transition from other pages
 while 1
   
   last_page = $page
@@ -53,22 +51,28 @@ while 1
 #    if Time.now - last_run > 6  
       page_out = threaded ? app.thread_wth_modules : app.run_wth_modules
 
-      app.reposition
-      app.cursor_hide
-      app.console_header(app.page_title($page - 1)).each {|o| app.clear_line; puts o; }
-   
       out = page_out[$page - 1] || ['Nothing to show']
 
-      out.each {|o|
-        app.clear_line
-        puts " #{o}"
-      }
-      app.clear_screen_down
+      if app.console_out
+        app.reposition
+        app.cursor_hide
+        app.console_header(app.page_title($page - 1)).each {|o| app.clear_line; puts o; }     
+        out.each {|o|
+          app.clear_line
+          puts " #{o}"
+        }
+        app.clear_screen_down
+      end
   
       app.webserver_pulse(page_out)
 #    end
   end    
 
-  24.times { break if app.keypress_pulse(0.25) }  
+  24.times {
+    if app.console_out
+      break if app.keypress_pulse
+    end
+    sleep(0.25)
+  }  
 end