+++ wthlab.rb
@@ -5,12 +5,12 @@
 # Love everybody,but never sell your sword.  ~ Paulo Coelho
 #
 require 'irbtools'
-load './core/common.rb'
+load './lib/common.rb'
 
 clear
 
 $app = app = Core.new(
     :config_file => "wth_config.yml"                 
 )
-
+puts "Use $app for access to WTH core app."
 binding.irb